Identity-ranking data with estimated weights based on inverse probability weighting

This data has two extra columns from the original `identity` data: `w` and `ranking`, which are respectively the estimated weights based on inverse probability weighting and the ranking pattern that the respondent provided, united into a single column.

Usage

identity_w

Format

## `identity_w` A data frame with 1,082 rows and 12 columns:

w

Estimated weight based on inverse probability weighting.

ranking

Ranking pattern that the respondent provided.

party

Marginal ranking for party (main identity ranking question).

religion

Marginal ranking for religion (main identity ranking question).

gender

Marginal ranking for gender (main identity ranking question).

race

Marginal ranking for race (main identity ranking question).

anc_house

Marginal ranking for household (anchor question).

anc_neighborhood

Marginal ranking for neighborhood (anchor question).

anc_city

Marginal ranking for city (anchor question).

anc_state

Marginal ranking for state (anchor question).

anc_correct_identity

Whether the respondent answered the anchor questions correctly. This is a binary variable that 1 if the respondent correctly answers the anchor ranking question and 0 if otherwise.

s_weight

Survey weight.

Source

<https://github.com/sysilviakim/ranking_error>